+++ src/SalesManager.java
@@ -15,9 +15,9 @@ public class SalesManager {
         return max;
     }
 
-    public int min() {
-        int min = sales[0];
-        for (int sale : sales) {
+    public long min() {
+        long min = sales[0];
+        for (long sale : sales) {
             if (sale < min) {
                 min = sale;
             }
@@ -25,9 +25,9 @@ public class SalesManager {
         return min;
     }
 
-    public int cutAverage() {
-        int sum = 0;
-        for (int sale : sales) {
+    public long cutAverage() {
+        long sum = 0;
+        for (long sale : sales) {
             sum += sale;
         }
         sum -= (max() + min());
